Lots of updates, see description
Completed the Billboard.CreateFromModel method, which now properly renders the model to a 2D image and then copies that image into the billboard object for future use

Allowed for cutout textures to be rendered. If R, G, B (which is normally between 0f and 1f) equals -1, -1, -1, then that pixel in the texture will be considered transparent, and it will not be rendered. There is no alpha, only cutout. Therefore, the cutouts may be abrupt and or jagged in some places.

Added a Utils.cpp and Utils.h which contain many convenience functions, including several print functions which print text to "output.txt" to help people with debugging

Removed the BillboardModel class, as it is still undecided how exactly we want to combine multiple billboards together

Added a function to the GzTextureStruct class called InitFromPPMFile which loads a PPM file into the texture